+++ packages/react-web-kit/src/components/molecules/Drawer/animations.ts
@@ -1,0 +1,61 @@
+export const defaultStyleOverlay = {
+  transition: `opacity 700ms ease-in-out`,
+  opacity: 0,
+};
+export const transitionStylesOverlay = {
+  entering: { opacity: 1 },
+  entered: { opacity: 1 },
+  exiting: { opacity: 0 },
+  exited: { opacity: 0 },
+};
+
+export const transformLeftRigthClose = (anchorPosition: string) =>
+  anchorPosition === 'right' ? 'translateX(100%)' : 'translateX(-100%)';
+
+export const transformLeftRigthOpen = 'translateX(0)';
+
+export const transformTopBottomClose = (anchorPosition: string) =>
+  anchorPosition === 'top' ? 'translateY(-100%)' : 'translateY(100%)';
+
+export const transformTopBottomOpen = 'translateY(0%)';
+
+export const defaultStylesLeftRight = (anchorPosition: string) => {
+  return {
+    transition: 'transform 300ms ease-in-out',
+    transform: () => transformLeftRigthClose(anchorPosition),
+    overflow: 'hidden',
+  };
+};
+
+export const transitionStylesLeftRight = (anchorPosition: string) => {
+  return {
+    entering: { transform: transformLeftRigthClose(anchorPosition) },
+    entered: {
+      transform: transformLeftRigthOpen,
+    },
+    exiting: {
+      transform: transformLeftRigthOpen,
+    },
+    exited: { transform: transformLeftRigthClose(anchorPosition) },
+  };
+};
+
+export const defaultStylesTopBottom = (anchorPosition: string) => {
+  return {
+    transition: 'transform 300ms ease-in-out',
+    transform: transformTopBottomClose(anchorPosition),
+    overflow: 'hidden',
+  };
+};
+export const transitionStylesTopBottom = (anchorPosition: string) => {
+  return {
+    entering: { transform: transformTopBottomClose(anchorPosition) },
+    entered: {
+      transform: transformTopBottomOpen,
+    },
+    exiting: {
+      transform: transformTopBottomOpen,
+    },
+    exited: { transform: transformTopBottomClose(anchorPosition) },
+  };
+};