+++ packages/react-web-kit/src/components/molecules/Select/SelectItem/index.ts
@@ -1,0 +1,1 @@
+export { default as ItemSelect } from './SelectItem';